+++ plugins/object.lua
@@ -9,7 +9,8 @@ local PREFIX = "[RMT_OBJ]"
 local callbacks = {
     Init = {},
     Step = {},
-    Draw = {}
+    Draw = {},
+    Hitbox = {}
 }
 
 
@@ -30,17 +31,44 @@ end
 
 
 Object.spawn = function(object, x, y)
-    local object = object - Object.ID_encoding
-
-    local drop = Item.spawn_drop(object, x, y, -4)
+    local drop = Item.spawn_drop(object - Object.ID_encoding, x, y, -4)
     drop.RMT_Object = object
 
     -- Run all Init callbacks on the object
-    for _, fn in ipairs(callbacks["Init"][object]) do
+    for _, fn in ipairs(callbacks["Init"][object - Object.ID_encoding]) do
         fn(drop)
     end
 
     return drop
+end
+
+
+Object.get_collisions = function(instance, index)
+    local cols = {}
+    local bbox = callbacks["Hitbox"][instance.RMT_Object - Object.ID_encoding]
+
+    for i = 0, gm.instance_number(index) - 1 do
+        local inst = gm.instance_find(index, i)
+        if Instance.exists(inst) then
+            if not (inst.bbox_left > instance.x + bbox.right or inst.bbox_right < instance.x + bbox.left
+                or inst.bbox_top > instance.y + bbox.bottom or inst.bbox_bottom < instance.y + bbox.top) then
+                table.insert(cols, inst)
+            end
+        end
+    end
+
+    return cols, #cols > 0
+end
+
+
+Object.get_collision_box = function(instance)
+    local hitbox = callbacks["Hitbox"][instance.RMT_Object - Object.ID_encoding]
+    return {
+        left    = hitbox.left   + instance.x,
+        top     = hitbox.top    + instance.y,
+        right   = hitbox.right  + instance.x,
+        bottom  = hitbox.bottom + instance.y
+    }
 end
 
 
@@ -57,10 +85,26 @@ Object.create = function(namespace, identifier)
     callbacks["Init"][object] = {}
     callbacks["Step"][object] = {}
     callbacks["Draw"][object] = {}
+    callbacks["Hitbox"][object] = {
+        left    = 0,
+        top     = 0,
+        right   = 0,
+        bottom  = 0
+    }
 
     -- Return object ID, with an encoding of +10000
     -- so that they remain separate from GM object_indexes
     return object + Object.ID_encoding
+end
+
+
+Object.set_hitbox = function(object, left, top, right, bottom)
+    callbacks["Hitbox"][object - Object.ID_encoding] = {
+        left    = left,
+        top     = top,
+        right   = right,
+        bottom  = bottom
+    }
 end
 
 
@@ -84,7 +128,7 @@ end
 
 gm.pre_code_execute(function(self, other, code, result, flags)
     if code.name:match("oCustomObject_pPickupItem_Step_0") and self.RMT_Object then
-        for _, fn in ipairs(callbacks["Step"][self.RMT_Object]) do
+        for _, fn in ipairs(callbacks["Step"][self.RMT_Object - Object.ID_encoding]) do
             fn(self)
         end
         return false
@@ -94,7 +138,7 @@ end)
 
 gm.pre_code_execute(function(self, other, code, result, flags)
     if code.name:match("oCustomObject_pPickupItem_Draw_0") and self.RMT_Object then
-        for _, fn in ipairs(callbacks["Draw"][self.RMT_Object]) do
+        for _, fn in ipairs(callbacks["Draw"][self.RMT_Object - Object.ID_encoding]) do
             fn(self)
         end
         return false